+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improve documentation
 
+### Deprecated
+
+- Deprecate `push_with_unchecked`, use `push_unchecked` instead
+
 ## [2.0.0] - 2026-02-09
 
 ### Added